Use double instead of float to allow big values of GameCycle (like on winchgate server)
--HG--
branch : fix_speed

+++ code/ryzom/server/src/entities_game_service/game_item_manager/game_item.h
@@ -651,9 +651,9 @@ public :
 	bool getStats(const std::string &stats, std::string &final );
 
 	/// accessors to the action latency end date
-	inline float getLatencyEndDate(){ return _LatencyEndDate; }
+	inline double getLatencyEndDate(){ return _LatencyEndDate; }
 
-	inline void setLatencyEndDate( float latencyEndDate ){ _LatencyEndDate = latencyEndDate; }
+	inline void setLatencyEndDate( double latencyEndDate ){ _LatencyEndDate = latencyEndDate; }
 
 	/// set the max sap load craft parameter
 	inline void setMaxSapLoad(float value)
@@ -950,7 +950,7 @@ private:
 	/// string associated with this item
 	std::string			_PhraseId;
 	/// tick when the proc will be available again
-	float	_LatencyEndDate;
+	double	_LatencyEndDate;
 	/// image of the item in bag / equipment
 //	uint16				_SlotImage;
 	NLMISC::TGameCycle	_TotalSaleCycle;
